proto: generate a serde-only json type tree
Add a gen-json feature exposing opentelemetry_proto::json, the same OTLP
messages as opentelemetry_proto::tonic with serde support and no prost
dependency. The tree is produced by the existing prost-build configuration
with client/server generation off and every prost reference stripped.

The serde helpers move to src/proto/serializers.rs and are mounted once per
tree so each tree serializes its own AnyValue type.
