+++ opentelemetry-proto/src/proto/json/opentelemetry.proto.collector.logs.v1.rs
@@ -1,0 +1,53 @@
+// This file is @generated by prost-build.
+#[derive(serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+#[derive(Clone, PartialEq, Debug, Default)]
+pub struct ExportLogsServiceRequest {
+    /// An array of ResourceLogs.
+    /// For data coming from a single resource this array will typically contain one
+    /// element. Intermediary nodes (such as OpenTelemetry Collector) that receive
+    /// data from multiple origins typically batch the data before forwarding further and
+    /// in that case this array will contain multiple elements.
+    pub resource_logs: ::std::vec::Vec<
+        super::super::super::logs::v1::ResourceLogs,
+    >,
+}
+#[derive(serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+#[derive(Clone, PartialEq, Eq, Hash, Debug, Default)]
+pub struct ExportLogsServiceResponse {
+    /// The details of a partially successful export request.
+    ///
+    /// If the request is only partially accepted
+    /// (i.e. when the server accepts only parts of the data and rejects the rest)
+    /// the server MUST initialize the `partial_success` field and MUST
+    /// set the `rejected_<signal>` with the number of items it rejected.
+    ///
+    /// Servers MAY also make use of the `partial_success` field to convey
+    /// warnings/suggestions to senders even when the request was fully accepted.
+    /// In such cases, the `rejected_<signal>` MUST have a value of `0` and
+    /// the `error_message` MUST be non-empty.
+    ///
+    /// A `partial_success` message with an empty value (rejected\_<signal> = 0 and
+    /// `error_message` = "") is equivalent to it not being set/present. Senders
+    /// SHOULD interpret it the same way as in the full success case.
+    pub partial_success: ::core::option::Option<ExportLogsPartialSuccess>,
+}
+#[derive(serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+#[derive(Clone, PartialEq, Eq, Hash, Debug, Default)]
+pub struct ExportLogsPartialSuccess {
+    /// The number of rejected log records.
+    ///
+    /// A `rejected_<signal>` field holding a `0` value indicates that the
+    /// request was fully accepted.
+    pub rejected_log_records: i64,
+    /// A developer-facing human-readable message in English. It should be used
+    /// either to explain why the server rejected parts of the data during a partial
+    /// success or to convey warnings/suggestions during a full success. The message
+    /// should offer guidance on how users can address such issues.
+    ///
+    /// error_message is an optional field. An error_message with an empty value
+    /// is equivalent to it not being set.
+    pub error_message: ::std::string::String,
+}
