+++ opentelemetry-proto/src/proto/json/opentelemetry.proto.resource.v1.rs
@@ -1,0 +1,22 @@
+// This file is @generated by prost-build.
+/// Resource information.
+#[derive(serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+#[serde(default)]
+#[derive(Clone, PartialEq, Debug, Default)]
+pub struct Resource {
+    /// Set of attributes that describe the resource.
+    /// Attribute keys MUST be unique (it is not allowed to have more than one
+    /// attribute with the same key).
+    /// The behavior of software that receives duplicated keys can be unpredictable.
+    pub attributes: ::std::vec::Vec<super::super::common::v1::KeyValue>,
+    /// The number of dropped attributes. If the value is 0, then
+    /// no attributes were dropped.
+    pub dropped_attributes_count: u32,
+    /// Set of entities that participate in this Resource.
+    ///
+    /// Note: keys in the references MUST exist in attributes of this message.
+    ///
+    /// Status: \[Development\]
+    pub entity_refs: ::std::vec::Vec<super::super::common::v1::EntityRef>,
+}
